docs: :art: Final reorganization of development documentation (#373)
* docs: :art: Final reorganization of development documentation

This PR concludes planned updates to the organization of development documentation. Additionally, I replaced the original table of contents using a tool that generates a nested list based on header levels, which I find easier to navigate when looking for specific details. I've applied the same generated table to all sub-pages with multiple headers for ease of navigation.

* docs: :memo: Remove cypress reference/point to playwright doc

The development documentation still had references to Cypress as the e2e testing tool. I removed the reference to cypress and linked out to the more extensive how-to guide for writing and running playwright tests.